src: move find_all_wakefiles() from runtime to frontend
